@@ -1,62 +1,58 @@
-// WeatherBanner Component
-
 import React from 'react';
 import PropTypes from 'prop-types';
 import moment from 'moment';
 import styled from 'styled-components';
 import iconCodeMapping from '../WeatherIcon';
 
-class WeatherBanner extends React.Component {
-  constructor(props) {
-    super(props);
-    this.data = props.data;
-    this.unit = props.unit;
-    this.locale = props.locale;
-  }
-
-  render() {
-    const data = this.data;
-    const unit = this.unit;
-    const locale = this.locale;
-    return (
-      <div>
-        <h5>{`${moment
-          .unix(data.dt)
-          .locale(locale)
-          .format('dddd a h:mm')}, ${data.desc}`}</h5>
-        <BannerContainer>
-          <BannerIcon src={iconCodeMapping[data.icon]} />
-          <Temperature>{Math.round(data.temp * 10) / 10}</Temperature>
-          <Unit>&deg;{unit === 'metric' ? 'C' : 'F'}</Unit>
-          <div style={{ flex: '1' }} />
-          <DetailContainer>
-            <InfoText>
-              Clouds: <b>{data.clouds}%</b>
-            </InfoText>
-            <InfoText>
-              Humidity: <b>{data.humidity}%</b>
-            </InfoText>
-            <InfoText>
-              Wind:{' '}
-              <b>
-                {data.wind}
-                {unit === 'metric' ? 'm/s' : 'mph'}
-              </b>
-            </InfoText>
-          </DetailContainer>
-        </BannerContainer>
-      </div>
-    );
-  }
-}
+/**
+ * Render a primary display of the current forecast, including a date time, a weather icon,
+ * current temperature, humidity, cloud density and wind
+ *
+ * @param {object} forecastNow the current forecast
+ * @param {string} unit the unit format for figures, only accepting 'metric' for now
+ * @param {locale} locale locale for time formating
+ */
+const WeatherBanner = ({ forecastNow, unit, locale }) => (
+  <div>
+    <h5>
+      ${moment.unix(forecastNow.dt).locale(locale).format('dddd a h:mm')}, $
+      {forecastNow.desc}
+    </h5>
+    <BannerContainer>
+      <BannerIcon src={iconCodeMapping[forecastNow.icon]} />
+      <Temperature>{Math.round(forecastNow.temp * 10) / 10}</Temperature>
+      <Unit>
+        &deg;
+        {unit === 'metric' ? 'C' : 'F'}
+      </Unit>
+      <div style={{ flex: '1' }} />
+      <DetailContainer>
+        <InfoText>
+          Clouds:
+          <b>{forecastNow.clouds}%</b>
+        </InfoText>
+        <InfoText>
+          Humidity: <b>{forecastNow.humidity}%</b>
+        </InfoText>
+        <InfoText>
+          Wind:{' '}
+          <b>
+            {forecastNow.wind}
+            {unit === 'metric' ? 'm/s' : 'mph'}
+          </b>
+        </InfoText>
+      </DetailContainer>
+    </BannerContainer>
+  </div>
+);
 
 WeatherBanner.defaultProps = {
   unit: 'metric',
-  locale: 'en-us',
+  locale: 'zh-tw',
 };
 
 WeatherBanner.propTypes = {
-  data: PropTypes.shape({
+  forecastNow: PropTypes.shape({
     dt: PropTypes.number.isRequired,
     temp: PropTypes.number.isRequired,
     temp_min: PropTypes.number.isRequired,
@@ -99,4 +95,6 @@ const DetailContainer = styled.div`
   flex-direction: column;
 `;
 
-const InfoText = styled.div`text-align: right;`;
+const InfoText = styled.div`
+  text-align: right;
+`;
